Move api helpers to the helpers fodler
